Add: Implemented the basic framework for Remote Admin authentication and storage. Change: Split the standard C++11's RNG object into it's own class separate from the Helper namespace. The only downside is we're required to create the object when needed per-class. Change: Begin work on updating the Player table with Connection Time and Packet I/O information.
Change: In the event a ':MIX3' or ':MIX7' packet is received we will attempt to check if the Player has already authenticated as a Remote Admin ( Supposing the Admin connected without a serNum ). Otherwise we check if the SerNum defined in these packets differs from the SerNum the Player class has on record for them and disconnect the User if it has changed.

Change: The Player class is now aware of the QItemIndex used by the playerTable. This makes updating table rows fairly straightforward.

TODO: Implement functions to programmatically create Remote Admins.
TODO: Implement Remote Admin Commands. All commands will be Rank-aware thus a level 0 'Game Master' will be unable to use level 3 'Owner' commands.