Change: Implemented some simple WoS-like baud measurements. Change: Re-implement sending packets to select Users / Scenes. ( Reduced the amount of redundant code. )
Change: Add a 100ms delay on disconnecting Users entering the wrong server password. ( This is so our 'incorrect' response will be sent. )
Change: Rename the "Admin" namespace to "AdminHelper". ( This is to reduce any confusion as to the purpose of the namespace. )

Change: The Player class will now update it's own Table row. ( This is to keep the Player class data from leaking into the ReMix class. )

Add: Begin implementing context menus for the PlayerTable.

TODO: Implement Admin commands.
TODO: Implement hooks to the 'BannedIP' and 'BannedSernum' classes for use with future Admin commands..