Change: Log Bans to file. ( ./banLog/DATE.txt ) Change: Add Helper to obtain a disconnect reason from the User.
Change: Fix a possible crash when deleting Players from the server.
Change: We now keep Users connected until the server attempts to update their status. This was done to allow the User to receive any disconnect or ban message the server may send.
Change: We now attempt to detect when a User is flooding the Server with packets. ( Limit within 2000MS is 256 packets. ( As defined by: PACKET_FLOOD_LIMIT. ) )
Change: Add Helper to send Master messages to select/all Users.
Change: Add function to send Server Rules as required.
Change: Add function to get a User's slot from their (QString)IP address.

Add: Partially implement a Packet decryption plugin (Commented out. ). The main use for this plugin is to allow the server to host multiple Worlds/Games.
Note: Due to the sensitivity of the plugin, it will not be distributed along with ReMix's Source. The only portion that can be distributed is the interface for this plugin as it is required. (Not included yet.)