FIX: Prevent private packets sent to Users and User Scenes from leaking to other Users.
Change: Change how QStrings are concocted by using QStringBuilder's '%' operator. This is to reduce the amount of allocations required to build the output QString.
Change: Add unused Helper to get a string from another string. (Untested function.)
Change: Auto-Default the Master-Checkin to every 2 seconds until a response is received from the Master. At which point the interval is set to every 5 minutes. (This is to reduce the instance of a ReMix server coming Online with no Master information.)