Change: Revert some SerNum-related variables to (qint32) casts as negative serNums are possible. Change: A User may now be both gracefully and ungracefully disconnected. The previous of the two will attempt to write everything within the buffer to the network before killing the connection while the latter will not.
Change: Properly check for Banned SerNums and invalid SerNums (If the server requires a SerNum) when the User fetches the Server's information, or connects via a Telnet service.
Change: Server Name information was reformatted slightly so as to remove code duplication.

Change: Implement the previously unused/untested function to get a sub-string from a string (Within the Helper namespace). The main reason for this is to prevent code duplication within the various classes/functions that require such a sub-string to be found and used.
Change: Re-implement command-line argument parsing. This should make the function easier to read and understand.
Change: Add a helper for converting a SerNum string (e.g. SOUL 4000 or 5DD423D1) to a qint32 (int) in a much cleaner manner than previously used. (Also clearly shows our intent)