Fix: Another bug with Helper::getStrStr. Previously the function would return an empty QString if the indStr was empty. (This is one weird function..)
Change: Implement code related to disconnecting IDLE Users. (Users that are connected, but haven't sent packets within ~10 minutes.)
Note: This does not supplement the actual Rule used via the User's Client and should only be used if you wish to keep Users relevant to each other (Active).

Change: Implement a new (two new..) context menu to Mute and Un-Mute a User's network. (The server will not re-send packets sent from this User to other users.)
Note: If packet decryption is enabled, it may be possible to mute only specific packets from a Muted User.

Change: Refactor code related to the context menu actions used by the ReMix class. This was done to reduce code duplication.
Change: The BanIP and BanSerNum prompts now inform the Host of the User's SerNum/IP when requesting confirmation.

Change: No longer offer the Option to close the Server if either the UDP or TCP socket was unable to successfully bind.
Change: Move some UDP-related code into the ServerInfo class. (This is mainly due to the fact the UDP socket relies on data from this class.)