Change: Remote-Admin Authentication will now be handled via a slot/signal within the User's UI update. Note: An Authentication event will only be attempted if the User has entered the Server's Password (If required.).
Change: The Server Host is now able to change the default host used to request the Master Server's location. (SynReal.ini).
Note: There are no checks for validity in place. As such, there is no way to default to the Syn-Real address.

Change: Remove some redundant checks related to the User's Socket.