Change: Convert parenthesis to square brackets within (most) messages.
FIX: An issue with refactoring the command handler led to Admin commands being available to all Users.
Note: To fix this we moved Admin-Checking code into a new function: CmdHandler::canUseAdminCommands(Player*)
Note 2: We only inform the User of command success when the User is able to use the command.