Change: Split the Settings GUI into a sub-widget. Change: Add Widget and Class related to setting Rules. (No code is implemented yet.)
Change: When toggling the password setting and a stored password is on file, check if the User would like to re-use it.
Fix: Set the 'pwdCheckState' boolean when the SettingsWidget object is first created.

Change: We're now setting the 'playerCount' within the ServerInfo class in preparation for server-side 'maxp' rule checking.