Note: These changes were made a very long time ago. I am currently unsure of the reasoning behind them, but I'll try to list what I can see.
Change: Disabled the Shutdown and Restart admin commands due to the current inability to shutdown or restart a server instance.
Change: Add the ability to start previously run servers from the stored preferences.
Change: Remove unused serverID references placed in multiple classes for use within the MultiServer implementation. (The logic ended up being changed to where they were not needed)

Change: Correctly mark past-servers as running or not running to prevent the same server from running twice.
Note: The above change has a side effect where if the server is incorrectly shut down the server will not be marked as shutdown.

Change: Remove the ability to change a server's name once it has been initialized.
Note: The above change was done to make saving and restoring server options easier.

Change: Correctly handle the closing of Server Instances and of closing the main process.

Fix: Remove all carriage returns and new-lines from messages sent to users via the admin dialog.
Note: the above fix was done to prevent only a portion of a message from being sent if it was copy and pasted into the dialog.