Add: Implement a rudimentary UPNP stack to enable Users to more easily use the application behind routers. Note: This has only been tested on my own local machine and router. As such, I cannot guarantee that the stack will correctly function.
Add: Users can now mark a local IP address as invalid from within the application by clicking on the server's IP address within the ServerStatus message.

Change: Changed how World names are shown to the user within the World Select dialog.
Change: Speed up UI information refresh rates to every half a second. This is to reduce the visible time for relevant information to be updated.

Change: Add a toggle to enable UPNP Port Forwarding adjacent to the "Public Server" toggle.
Note: It is currently labeled "!" due to space constraints, but will receive a proper name and location later on.

Fix: Properly close the MasterSocket before rebinding it to a new local IP address.
Fix: Properly close and delete the UPNP Object when the User disables the setting.