Fix: I've modified all "QObject::Connect" lambdas to be more aware of when their sending object has been deleted. This should prevent any possible crashes related to an Object's lambda being called while the Object is deleted.
Change: Sending a message to a player from the ReMix UI will now be done from within the Player class itself, and will also allow for sending messages to multiple individuals at once, or to all connected Players.
Note: This was done to replace a temporary fix for a crash related to sending messages to Disconnected Players.