Change: Use the Player Class to send all Player specific Messages from the ReMix Server. Change: Remove some verbosity from various Boolean checks.
Change: Removed the Banished Message from the MessagesWidget class.
Note: This was done due to the message never actually being sent due to the Player being disconnected when actually checking if the Player is banned.

Change: Rename the MessagesWidget  Class to "MOTDWidget" to better reflect it's usage.

Revert: Reverted a small portion of code related to sending Messages to Players in a commented form.