Change: Update Version Number to 2.1.6. Change: Apply some information related to the application via a MOC.
Fix: Properly handle changing a server from public to private and back in a 'quick' timeframe.
Note: Before, the Server would get confused as to which state it was supposed to be in and no longer have the ability to become public.

Fix: When starting a Server Instance as Public, the TCP Server will now initialize after being set to the ServerInfo class.
Note: Before, the Server(s) would show up on the Master List but were unable to be connected to.